+++ discO/plugin/agama/tests/test_wrapper.py
@@ -64,19 +64,19 @@ class Test_AGAMAPotentialWrapperMeta(
 
     # /def
 
-    def test_specific_potential(self):
+    def test_potential(self):
         """Test method ``specific_force``."""
         # ---------------
         # when there isn't a frame
 
         with pytest.raises(TypeError) as e:
-            self.subclass.specific_potential(self.potential, self.points)
+            self.subclass.potential(self.potential, self.points)
 
         assert "the potential must have a frame." in str(e.value)
         # ---------------
         # basic
 
-        points, values = self.subclass.specific_potential(
+        points, values = self.subclass.potential(
             self.potential,
             self.points.data,
         )
@@ -101,7 +101,7 @@ class Test_AGAMAPotentialWrapperMeta(
             "galactocentric",
         ):
 
-            points, values = self.subclass.specific_potential(
+            points, values = self.subclass.potential(
                 self.potential,
                 self.points,
                 frame=frame,
@@ -116,7 +116,7 @@ class Test_AGAMAPotentialWrapperMeta(
         # ---------------
         # representation_type
 
-        points, values = self.subclass.specific_potential(
+        points, values = self.subclass.potential(
             self.potential,
             self.points,
             frame=self.points.frame.replicate_without_data(),
@@ -248,12 +248,12 @@ class Test_AGAMAPotentialWrapper(
 
     # /def
 
-    def test_specific_potential(self):
-        """Test method ``specific_potential``."""
+    def test_potential(self):
+        """Test method ``potential``."""
         # ---------------
         # basic
 
-        points, values = self.inst.specific_potential(self.points.data)
+        points, values = self.inst.potential(self.points.data)
 
         # check data types
         assert isinstance(points, self.inst.frame.__class__)
@@ -266,7 +266,7 @@ class Test_AGAMAPotentialWrapper(
         # ---------------
         # with a frame
 
-        points, values = self.inst.specific_potential(self.points)
+        points, values = self.inst.potential(self.points)
 
         # check data types
         assert isinstance(points, coord.SkyCoord)
@@ -283,7 +283,7 @@ class Test_AGAMAPotentialWrapper(
 
         with pytest.raises(TypeError) as e:
 
-            points, values = self.inst.specific_potential(
+            points, values = self.inst.potential(
                 self.points,
                 frame=coord.Galactocentric(),
             )
@@ -293,7 +293,7 @@ class Test_AGAMAPotentialWrapper(
         # ---------------
         # representation_type
 
-        points, values = self.inst.specific_potential(
+        points, values = self.inst.potential(
             self.points,
             representation_type=coord.CartesianRepresentation,
         )